+++ hook/server.go
@@ -4,6 +4,8 @@ import (
 	"context"
 	"fmt"
 	"net/http"
+
+	"github.com/golang/glog"
 )
 
 //go:generate counterfeiter -o ../mocks/manager.go --fake-name Manager . manager
@@ -16,9 +18,12 @@ type Server struct {
 }
 
 func (s *Server) ServeHTTP(resp http.ResponseWriter, req *http.Request) {
+	glog.V(1).Info("sync changes triggerd")
 	if err := s.Manager.Run(req.Context()); err != nil {
+		glog.Warningf("sync changes failed: %v", err)
 		http.Error(resp, fmt.Sprintf("run failed: %s", err.Error()), http.StatusInternalServerError)
 		return
 	}
+	glog.V(1).Info("sync changes completed successful")
 	fmt.Fprintln(resp, "ok")
 }